requirements.txt
python-dotenv
flask
requests
bcrypt
pymongo
redis
flask_session
flask_limiter
sib-api-v3-sdk
web3
cryptography
